requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
tiktoken
python-dotenv
chromadb
google-cloud-firestore
faiss-cpu